requirements.txt
asgiref==3.5.2
defusedxml==0.7.1
diff-match-patch==20200713
Django==4.1.3
django-admin-list-filters==1.1
django-admin-thumbnails==0.2.6
django-import-export==3.0.1
et-xmlfile==1.1.0
MarkupPy==1.14
odfpy==1.4.1
openpyxl==3.0.10
Pillow==9.2.0
PyYAML==6.0
six==1.16.0
sqlparse==0.4.3
tablib==3.2.1
xlrd==2.0.1
xlwt==1.3.0
